// src/frontend-applications/agent-portal/modules/cmdb/model/ConfigItemClassProperty.ts
/**
 * Copyright (C) 2006-2024 KIX Service Software GmbH, https://www.kixdesk.com
 * --
 * This software comes with ABSOLUTELY NO WARRANTY. For details, see
 * the enclosed file LICENSE for license information (GPL3). If you
 * did not receive this file, see https://www.gnu.org/licenses/gpl-3.0.txt.
 * --
 */

export enum ConfigItemClassProperty {

    ID = 'ID',

    NAME = 'Name',

    VALID_ID = 'ValidID',

    COMMENT = 'Comment',

    CHANGE_TIME = 'ChangeTime',

    CHANGE_BY = 'ChangeBy',

    CREATE_TIME = 'CreateTime',

    CREATE_BY = 'CreateBy',

    DEFINITIONS = 'Definitions',

    CURRENT_DEFINITION = 'CurrentDefinition',

    DEFINITION_STRING = 'DefinitionString',

    ICON = 'ICON'

}